openplains_api/requirements.txt
Django>=4.0.5
psycopg2-binary>=2.8
djangorestframework>=3.13
django-cors-headers>=3.11
django-environ>=0.8
markdown>=3.3
django-filter>=21.1
django-guardian>=2.4.0
django-debug-toolbar>=3.2
django-extensions>=3.1
djangorestframework-gis>=0.18
django-rest-knox>=4.2.0
redis>=4.0
hiredis>=2.0
requests>=2.27
django-storages[google]>=1.12
pillow>=9.0
channels>=3.0
channels_redis>=3.4
selenium>=4.1.3
celery[redis]>=5.2.3
jupyter>=1.0.0
notebook>=6.4.10
pyparsing>=3.0.8
pydot>=1.4.2
Werkzeug>=2.1.1
pystac-client>=0.5.0